System Information utility using C#
Your final project is to implement a System Information utility using C#.  You will create a form and a series of command buttons to call up to the screen the system information for the categories of hardware listed below.  You can use the sample application from class as the starting point for the app. 
Show basic and advanced hardware and software information for the following hardware categories.  Display the following hardware information:
GPU (Video Controller Class), Hard Drive, Processor, OS, Printers, Network Adapters, Cache Memory, Device Bus, Disk Drive, Disk Partition, IDE Controller, IRQ Resources, Sound Device.
Format all memory numbers in Megabytes (MB, GB, TB).  So if a value is 500K bytes it will be .5 MB.  If it is 123,456,789,000 the value will be 123.4 GB.  123,456,789,000,000 will be 123.4 TB, etc.   
In addition, implement a File and directory viewer.  Start at the root C:\ and then allow the user to enter a directory structure and then display the contents of the directory and the file attributes like file name, date created, last modified, file size.
The last requirement for this assignment is to give the user the ability to output the information in the display to a text file.  When the user clicks the Save to File button, utilize the file save dialog control to allow the user to navigate to a folder and enter a file name.  Save all of the information in the display to this file.  The text file format is fine for this assignment.  Call the output file Output.txt.
Extra Credit
For 10 points extra credit adds a File utility to copy, move and delete files.  Pull up a window list or combo box starting at C:\ root directory, be able to navigate to folders and files and copy, move and delete these files.